AI_travel_planner/requirements.txt
flask
python-dotenv
openai>=1.44.0
streamlit